Massive 2.0 commit
See Lux-EA-B2 release for description and changes.

+++ shaders/lib/util/framebufferFormats.glsl
@@ -8,7 +8,7 @@ See AGREEMENT.txt for more information.
 
 /*
 const int colortex0Format   = R11F_G11F_B10F;     // Main
-const int colortex1Format   = RGB8;               // Raw Translucent, Bloom
+const int colortex1Format   = RGB16F;             // Raw Translucent, Bloom
 const int colortex2Format   = RGBA16;             // Temporal Stuff
 const int colortex3Format   = RGB8;               // Specular Data
 const int gaux1Format       = R8;                 // Cloud Alpha
